put symbol line function in separate file
