+++ screens/status_system.js
@@ -17,8 +17,8 @@ var round = require('../lib/utils/utils.js').round;
 var convertBg= require('../lib/utils/utils.js').convertBg;
 var stripLeadingZero= require('../lib/utils/utils.js').stripLeadingZero;
 
-var drawReservoirIcon = require('../lib/utils/utils.js').drawReservoirIcon;
-var drawBatteryIcon = require('../lib/utils/utils.js').drawBatteryIcon;
+var drawSymbolLine = require('./symbol_line.js');
+
 var drawWiFiIcon = require('../lib/utils/utils.js').drawWiFiIcon;
 var drawBTIcon = require('../lib/utils/utils.js').drawBTIcon;
 var drawConnectIcon = require('../lib/utils/utils.js').drawConnectIcon;
@@ -29,7 +29,7 @@ const execSync = require('child_process').execSync;
 //
 
 // wipe bottom line for error messages if already written
-function wipeIfAlreadyUsed(yOffset){
+function wipeIfAlreadyUsed(display, yOffset){
   if (yOffset >= 56){
     yOffset = 56;
     display.oled.fillRect(0, yOffset, 128, 10,  0, false);  
@@ -38,101 +38,12 @@ function wipeIfAlreadyUsed(yOffset){
 }
 
 module.exports = systemStatus;
-function systemStatus(display, openapsDir) {
+function systemStatus(display, openapsDir, pumpPref) {
 
 //clear display buffer
 display.oled.clearDisplay(true); 
 
-//
-// BEGIN Symbol Line (from left to right)
-//
-try {
-    var pumpBatterylevel = JSON.parse(fs.readFileSync(openapsDir+"/monitor/battery.json"));
-} catch (e) {
-    console.error("Status screen display error: could not parse battery.json: ", e);
-}
-try {
-    var reservoir = JSON.parse(fs.readFileSync(openapsDir+"/monitor/reservoir.json"));
-} catch (e) {
-    console.error("Status screen display error: could not parse reservoir.json: ", e);
-}
-try{
-	var hasPublicIp = fs.existsSync('/tmp/hasPublicIp');
-} catch (e) {
-	// not online
-  console.log('No "/tmp/hasPublicIp" found. Not online?');
-}
-try {
-	var wifiIp = execSync('ip -f inet -o addr show wlan0|cut -d " " -f 7 |cut -d "/" -f 1').toString().trim();
-} catch (e){
-	console.error("Status screen display error: could not execute wifiIp: ", e);
-}
-try {
-	var btIp = execSync('ip -f inet -o addr show bnep0|cut -d " " -f 7 |cut -d "/" -f 1').toString().trim();
-} catch (e){
-	console.error("Status screen display error: could not execute btIp: ", e);
-}
-try {
-    var batterylevel = JSON.parse(fs.readFileSync(openapsDir+"/monitor/edison-battery.json"));
-} catch (e) {
-    console.error("Status screen display error: could not parse edison-battery.json: ", e);
-}
-
-// show pump battery level
-if(pumpBatterylevel && pumpBatterylevel.voltage) { 
-  // set your battery voltage here
-  var voltageHigh = 1.7;
-  var voltageLow = 1.4;
-  
-  var battlevel = ((pumpBatterylevel.voltage - voltageLow) / (voltageHigh - voltageLow)) * 100.0;
-  battlevel = (battlevel > 100 ? 100 : battlevel);    
-  drawBatteryIcon(display, 0, 0 ,battlevel);
-} else {
-  drawBatteryIcon(display, 0, 0 ,-1);
-}
-
-// show pump reservoir icon
-if (reservoir){
-  drawReservoirIcon(display, 22, 0, reservoir);
-} else {
-  drawReservoirIcon(display, 22, 0, -1);
-}
-
-// show current time
-var nowDate = new Date();
-var hour = nowDate.getHours();
-hour = (hour < 10 ? "0" : "") + hour;
-var min  = nowDate.getMinutes();
-min = (min < 10 ? "0" : "") + min;
-
-display.oled.setCursor(50,1);
-display.oled.writeString(font, 1, hour+":"+min, 1, false, 0, false);
-
-
-// show online connection icon if connected to the Internet
-// TODO or WiFi AP icon if AP mode active
-if (hasPublicIp){
-  drawConnectIcon(display, 82, 1, true);
-} else {
-  drawConnectIcon(display, 82, 1, false);
-}
-
-// show WiFi icon if connected to a Wifi network or Bluetooth icon if connected to a PAN
-if (wifiIp){
-  drawWiFiIcon(display, 98, 0);
-} else if (btIp){
-  drawBTIcon(display, 99, 0);
-}
-
-// show local battery level
-if(batterylevel) {
-  drawBatteryIcon(display, 113, 0 , batterylevel.battery);
-} else {
-  drawBatteryIcon(display, 113, 0 ,-1);
-}
-//
-// END Symbol Line
-//
+drawSymbolLine(display, openapsDir, pumpPref);
 
 //
 // BEGIN System Status
@@ -143,9 +54,25 @@ try {
 	console.error("Status screen display error: could not execute hostname: ", e);
 }
 try {
+	var wifiIp = execSync('ip -f inet -o addr show wlan0|cut -d " " -f 7 |cut -d "/" -f 1').toString();
+} catch (e){
+	console.error("Status screen display error: could not execute wifiIp: ", e);
+}
+try {
+	var btIp = execSync('ip -f inet -o addr show bnep0|cut -d " " -f 7 |cut -d "/" -f 1').toString();
+} catch (e){
+	console.error("Status screen display error: could not execute btIp: ", e);
+}
+try {
 	var wifiName = execSync('iwgetid -r').toString().trim();
 } catch (e){
 	console.error("Status screen display error: could not execute wifiName: ", e);
+}
+try{
+	var hasPublicIp = fs.existsSync('/tmp/hasPublicIp');
+} catch (e) {
+	// not online
+  console.log('No "/tmp/hasPublicIp" found. Not online?');
 }
 try {
   if (hasPublicIp) 
@@ -242,27 +169,27 @@ if (yOffset === 16){
 if (status && suggested) {
     var notLoopingReason = suggested.reason;
     if (status.suspended == true) {
-        wipeIfAlreadyUsed(yOffset);
+        wipeIfAlreadyUsed(display, yOffset);
         display.oled.writeString(font, 1, "PUMP SUSPENDED", 1, false, 0, false);
     }
     else if (notLoopingReason.includes("CGM is calibrating")) {
-        wipeIfAlreadyUsed(yOffset);
+        wipeIfAlreadyUsed(display, yOffset);
         display.oled.writeString(font, 1, "CGM calib./???/noisy", 1, false, 0, false);
     }
     else if (notLoopingReason.includes("CGM data is unchanged")) {
-        wipeIfAlreadyUsed(yOffset);
+        wipeIfAlreadyUsed(display, yOffset);
         display.oled.writeString(font, 1, "CGM data unchanged", 1, false, 0, false);
     }
     else if (notLoopingReason.includes("BG data is too old")) {
-        wipeIfAlreadyUsed(yOffset);
+        wipeIfAlreadyUsed(display, yOffset);
         display.oled.writeString(font, 1, "BG data too old", 1, false, 0, false);
     }
     else if (notLoopingReason.includes("currenttemp rate")) {
-        wipeIfAlreadyUsed(yOffset);
+        wipeIfAlreadyUsed(display, yOffset);
         display.oled.writeString(font, 1, "Temp. mismatch", 1, false, 0, false);
     }
     else if (suggested.carbsReq) {
-        wipeIfAlreadyUsed(yOffset);
+        wipeIfAlreadyUsed(display, yOffset);
         display.oled.writeString(font, 1, "Carbs Required: "+suggested.carbsReq+'g', 1, false, 0, false);
     }
 //add more on-screen warnings/messages, maybe some special ones for xdrip-js users?
