+++ screens/status_text.js
@@ -15,129 +15,23 @@ var font = require('oled-font-5x7');
 var round = require('../lib/utils/utils.js').round;
 var convertBg= require('../lib/utils/utils.js').convertBg;
 
-var drawReservoirIcon = require('../lib/utils/utils.js').drawReservoirIcon;
-var drawBatteryIcon = require('../lib/utils/utils.js').drawBatteryIcon;
-var drawWiFiIcon = require('../lib/utils/utils.js').drawWiFiIcon;
-var drawBTIcon = require('../lib/utils/utils.js').drawBTIcon;
-var drawConnectIcon = require('../lib/utils/utils.js').drawConnectIcon;
+var drawSymbolLine = require('./symbol_line.js');
 
 var drawArrowUp = require('../lib/utils/utils.js').drawArrowUp;
 var drawArrowDown = require('../lib/utils/utils.js').drawArrowDown;
 var drawLoopIcon = require('../lib/utils/utils.js').drawLoopIcon;
 var drawTargetIcon = require('../lib/utils/utils.js').drawTargetIcon;
-const execSync = require('child_process').execSync;
-
-
-function drawSymbolLine(display, openapsDir){
-//
-// BEGIN Symbol Line (from left to right)
-//
-try {
-    var pumpPref = JSON.parse(fs.readFileSync("./config/pump.json"));
-} catch (e) {
-    console.error("Status screen display error: could not parse config/pump.json: ", e);
-    //var pumpPref = {pumpReservoirSize = 3, pumpBatteryTypes}
-}
-try {
-    var pumpBatterylevel = JSON.parse(fs.readFileSync(openapsDir+"/monitor/battery.json"));
-} catch (e) {
-    console.error("Status screen display error: could not parse battery.json: ", e);
-}
-try {
-    var reservoir = JSON.parse(fs.readFileSync(openapsDir+"/monitor/reservoir.json"));
-} catch (e) {
-    console.error("Status screen display error: could not parse reservoir.json: ", e);
-}
-try{
-	var hasPublicIp = fs.existsSync('/tmp/hasPublicIp');
-} catch (e) {
-	// not online
-  console.log('No "/tmp/hasPublicIp" found. Not online?');
-}
-try {
-	var wifiIp = execSync('ip -f inet -o addr show wlan0|cut -d " " -f 7 |cut -d "/" -f 1').toString();
-} catch (e){
-	console.error("Status screen display error: could not execute wifiIp: ", e);
-}
-try {
-	var btIp = execSync('ip -f inet -o addr show bnep0|cut -d " " -f 7 |cut -d "/" -f 1').toString();
-} catch (e){
-	console.error("Status screen display error: could not execute btIp: ", e);
-}
-try {
-    var batterylevel = JSON.parse(fs.readFileSync(openapsDir+"/monitor/edison-battery.json"));
-} catch (e) {
-    console.error("Status screen display error: could not parse edison-battery.json: ", e);
-}
-
-// show pump battery level
-if(pumpPref && pumpBatterylevel && pumpBatterylevel.voltage) { 
-  // set your battery voltage here
-  var voltageHigh = pumpPref.pumpBatteryTypes[pumpPref.pumpBatteryIndex].high;
-  var voltageLow = pumpPref.pumpBatteryTypes[pumpPref.pumpBatteryIndex].low;
-  
-  var battlevel = ((pumpBatterylevel.voltage - voltageLow) / (voltageHigh - voltageLow)) * 100.0;
-  battlevel = (battlevel > 100 ? 100 : battlevel);    
-  drawBatteryIcon(display, 0, 0 ,battlevel);
-} else {
-  drawBatteryIcon(display, 0, 0 ,-1);
-}
-
-// show pump reservoir icon
-if (pumpPref && reservoir){
-  reservoir = reservoir / (pumpPref.pumpReservoirSize / 100);
-  drawReservoirIcon(display, 22, 0, reservoir);
-} else {
-  drawReservoirIcon(display, 22, 0, -1);
-}
-
-// show current time
-var nowDate = new Date();
-var hour = nowDate.getHours();
-hour = (hour < 10 ? "0" : "") + hour;
-var min  = nowDate.getMinutes();
-min = (min < 10 ? "0" : "") + min;
-
-display.oled.setCursor(48,1);
-display.oled.writeString(font, 1, hour+":"+min, 1, false, 0, false);
-
-
-// show online connection icon if connected to the Internet
-// TODO or WiFi AP icon if AP mode active
-if (hasPublicIp){
-  drawConnectIcon(display, 82, 1, true);
-} else {
-  drawConnectIcon(display, 82, 1, false);
-}
-
-// show WiFi icon if connected to a Wifi network or Bluetooth icon if connected to a PAN
-if (wifiIp){
-  drawWiFiIcon(display, 98, 0);
-} else if (btIp){
-  drawBTIcon(display, 99, 0);
-}
-
-// show local battery level
-if(batterylevel) {
-  drawBatteryIcon(display, 113, 0 , batterylevel.battery);
-} else {
-  drawBatteryIcon(display, 113, 0 ,-1);
-}
-//
-// END Symbol Line
-//
-}
 
 //
 //Start of status display function
 //
 
 module.exports = textStatus;
-function textStatus(display, openapsDir) {
+function textStatus(display, openapsDir, pumpPref) {
 //clear the buffer
 display.oled.clearDisplay(true);
 
-drawSymbolLine(display, openapsDir);
+drawSymbolLine(display, openapsDir, pumpPref);
 
 //
 // BEGIN Text Status
@@ -221,10 +115,9 @@ if(bg && profile) {
       display.oled.fillRect(0, 25, curPos, 3, 1, false);
       
       display.oled.setCursor(curPos+5,16);
-      display.oled.writeString(font, 1, ""+minutes, 1, false, 0, false);
-      
-	    curPos = display.oled.cursor_x + 1 ;
-      display.oled.fillRect(curPos, 15, 1, 3, 1, false);
+      display.oled.writeString(font, 1, minutes+"'", 1, false, 0, false);      
+	    //curPos = display.oled.cursor_x + 1 ;
+      //display.oled.fillRect(curPos, 15, 1, 3, 1, false);
     } else if (delta) {
       if (delta >= 5) {
         drawArrowUp(display, curPos+3, 18);
